+++ config/packages/messenger.yaml
@@ -12,4 +12,4 @@ framework:
     routing:
       # Route your messages to the transports
       # 'App\Message\YourMessage': async
-      'App\Domain\Segment\Segment': async
+      'App\Segment\Domain\Segment': async
