+++ config/routes/annotations.yaml
@@ -1,3 +1,3 @@
 controllers:
-    resource: ../../src/Infrastructure/Http/Controller/
+    resource: ../../src/Segment/Http/Controller/
     type: annotation
