+++ src/Segment/Domain/Point.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 
 
-namespace App\Domain\Segment;
+namespace App\Segment\Domain;
 
 
 /**
