+++ src/Segment/Http/Controller/TaskController.php
@@ -1,12 +1,11 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Infrastructure\Http\Controller;
+namespace App\Segment\Http\Controller;
 
 
 use App\Infrastructure\Http\Response\Responder;
-use App\Infrastructure\Task\TaskRepository;
-use Doctrine\DBAL\DBALException;
+use App\Segment\Task\TaskRepository;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
@@ -37,7 +36,6 @@ final class TaskController
      * @param TaskRepository $taskRepository
      * @Route("/{token}", name="tasks_token_status", methods={"GET"})
      * @return Response
-     * @throws DBALException
      */
     public function status(
         string $token,
